Dev (#4)
* test: remove any archive folder from tests

* refactor: archive unused hourly-forecast-graph component

* chore: remove unused clsx import

* refactor: move current weather card component to new file

* chore: move unused imports from refactoring

Removes the unused imports left after refactoring the CurrentWeatherCard component to a new file.

* refactor: move day-weather-component to new file

- Refactor the code to a new file
- Renames the component from `DayComponent` to `DayWeatherComponent`

* refactor: ensure coherent filename casing

Rename the Error and Loading state components to align with the coherent casing throughout the project

* fix: wrong casing in imports

* fix: convert time to target instead of locale

Instead of having converting utc dates to the clients locale time, it instead converts it to the targeted location. This makes the datetime the same for everyone regardless of their local timezone.

Note: Right now only one target timezone is available "Europe/Stockholm".

* style(useWeather.ts): add new empty line